backend/pyproject.toml
[project]
name = "voice-bridge-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.0",
    "pydantic-settings>=2.6",
    "redis[hiredis]>=5.2",
    "google-genai>=1.10.0",
    "twilio>=9.3",
    "audioop-lts>=0.2.1",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "fakeredis>=2.26",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
